integration: Move CmdCommand to its own file
